refactor: move DeviceBackend to kvbm-common
promote the DeviceBackend enum from kvbm-physical to kvbm-common so
both the storage layer dynamo-memory and the device layer
kvbm-physical can reference a single canonical tag. Delete local copy
DeviceBackendKind that dynamo-memory::numa had been carrying.
